Stopping service in preremove was interfering with restart on upgrade

+++ pkg/scripts/preremove.sh
@@ -2,4 +2,12 @@
 
 SERVICE=telemetry-envoy
 
-systemctl stop ${SERVICE}.service
+
+# debs pass upgrade, remove, or purge
+# rpms pass number of versions installed, so 0 means none remain
+if [ "$1" == "remove" ] || [ "$1" == "purge" ] || [ "$1" == "0" ]; then
+
+  systemctl stop ${SERVICE}.service
+
+fi
+
